Split ingestion integration tests
Move end-to-end multi-source ingestion scenarios into a focused test module
and extract shared provenance and reference snapshot helpers into support code.
Remove the file-level `too-many-lines` suppression from the validation tests.
