+++ tests/test_ingestion_integration.py
@@ -1,31 +1,12 @@
-"""Integration tests for the multi-source ingestion service."""
+"""Validation tests for the multi-source ingestion service."""
 
-# End-to-end ingestion scenarios share setup helpers and assertions in one file.
-# pylint: disable=too-many-lines
-
-import dataclasses
-import datetime as dt
 import typing as typ
-import uuid
 
 import pytest
-from _ingestion_service_helpers import (
-    _get_job_record_for_episode,
-    _make_raw_source,
-)
+from _ingestion_service_helpers import _make_raw_source
 
-from episodic.canonical.domain import (
-    ReferenceBinding,
-    ReferenceBindingTargetKind,
-    ReferenceDocument,
-    ReferenceDocumentKind,
-    ReferenceDocumentLifecycleState,
-    ReferenceDocumentRevision,
-)
+import tests.test_ingestion_integration_support as ingestion_support
 from episodic.canonical.ingestion import MultiSourceRequest
-from episodic.canonical.ingestion_service import ingest_multi_source
-from episodic.canonical.storage import SqlAlchemyUnitOfWork
-from tests.test_uuid_assertions import assert_uuid7
 
 if typ.TYPE_CHECKING:
     from sqlalchemy.ext.asyncio import AsyncSession
@@ -34,150 +15,10 @@ if typ.TYPE_CHECKING:
     from episodic.canonical.ingestion_service import IngestionPipeline
 
 
-class SourcePriorityRecord(typ.TypedDict):
-    """Serialized source-priority record in TEI provenance metadata."""
-
-    priority: int
-    source_uri: str
-    source_type: str
-    weight: float
-    content_hash: str
-
-
-class TeiHeaderProvenanceRecord(typ.TypedDict):
-    """Serialized TEI header provenance metadata payload."""
-
-    capture_context: str
-    ingestion_timestamp: str
-    source_priorities: list[SourcePriorityRecord]
-    reviewer_identities: list[str]
-
-
-@dataclasses.dataclass(frozen=True)
-class IngestionTestContext:
-    """Test fixtures for multi-source ingestion error tests."""
-
-    session_factory: typ.Callable[[], AsyncSession]
-    profile: SeriesProfile
-    ingestion_pipeline: IngestionPipeline
-
-
-def _require_provenance_payload(
-    payload: dict[str, object],
-) -> TeiHeaderProvenanceRecord:
-    """Return the TEI header provenance payload with runtime checks."""
-    provenance = payload.get("episodic_provenance")
-    assert isinstance(provenance, dict), (
-        "Expected TEI header payload to include dict provenance metadata."
-    )
-    return typ.cast("TeiHeaderProvenanceRecord", provenance)
-
-
 def test_raw_source_helper_rejects_unknown_overrides() -> None:
     """Raw source helper should reject typoed override keys."""
     with pytest.raises(ValueError, match="Invalid override keys"):
         _make_raw_source(**typ.cast("typ.Any", {"unknown": "value"}))
-
-
-def _verify_provenance_metadata(
-    provenance: TeiHeaderProvenanceRecord,
-    expected_reviewer: str,
-    expected_source_uris: list[str],
-) -> None:
-    """Verify provenance metadata values for an ingested TEI header."""
-    timestamp = provenance.get("ingestion_timestamp")
-    assert isinstance(timestamp, str), "Expected ingestion timestamp as string."
-    assert dt.datetime.fromisoformat(timestamp).tzinfo is not None, (
-        "Expected timezone-aware ingestion timestamp."
-    )
-    assert provenance.get("reviewer_identities") == [expected_reviewer], (
-        "Expected reviewer identity to be captured from request actor."
-    )
-    priorities = provenance["source_priorities"]
-    assert len(priorities) == len(expected_source_uris), (
-        "Expected one priority entry per source."
-    )
-    actual_source_uris = [priority["source_uri"] for priority in priorities]
-    assert actual_source_uris == expected_source_uris, (
-        "Expected source priorities to match source URI order."
-    )
-
-
-def _verify_source_documents(
-    documents: list[typ.Any],
-    expected_count: int,
-) -> None:
-    """Verify persisted source document count and weight bounds."""
-    assert len(documents) == expected_count, (
-        "Expected all input sources to be persisted as source documents."
-    )
-    for document in documents:
-        assert_uuid7(document.id, "source document")
-        assert document.weight > 0.0, (
-            "Expected persisted source weight to be greater than zero."
-        )
-        assert document.weight <= 1.0, (
-            "Expected persisted source weight to be capped at one."
-        )
-
-
-async def _create_reference_fixtures(
-    session_factory: typ.Callable[[], AsyncSession],
-    profile: SeriesProfile,
-) -> tuple[ReferenceDocument, ReferenceDocumentRevision, ReferenceBinding]:
-    """Persist the reference document fixtures used by snapshotting tests."""
-    now = dt.datetime.now(dt.UTC)
-    reference_document = ReferenceDocument(
-        id=uuid.uuid4(),
-        owner_series_profile_id=profile.id,
-        kind=ReferenceDocumentKind.STYLE_GUIDE,
-        lifecycle_state=ReferenceDocumentLifecycleState.ACTIVE,
-        metadata={"name": "Ingestion style guide"},
-        created_at=now,
-        updated_at=now,
-    )
-    reference_revision = ReferenceDocumentRevision(
-        id=uuid.uuid4(),
-        reference_document_id=reference_document.id,
-        content={"summary": "Use short declarative sentences."},
-        content_hash="ingestion-reference-hash",
-        author="editor@example.com",
-        change_note="Initial guidance",
-        created_at=now,
-    )
-    reference_binding = ReferenceBinding(
-        id=uuid.uuid4(),
-        reference_document_revision_id=reference_revision.id,
-        target_kind=ReferenceBindingTargetKind.SERIES_PROFILE,
-        series_profile_id=profile.id,
-        episode_template_id=None,
-        ingestion_job_id=None,
-        effective_from_episode_id=None,
-        created_at=now,
-    )
-    async with SqlAlchemyUnitOfWork(session_factory) as uow:
-        await uow.reference_documents.add(reference_document)
-        await uow.reference_document_revisions.add(reference_revision)
-        await uow.flush()
-        await uow.reference_bindings.add(reference_binding)
-        await uow.commit()
-    return reference_document, reference_revision, reference_binding
-
-
-async def _assert_ingestion_raises(
-    context: IngestionTestContext,
-    request: MultiSourceRequest,
-    expected_error_pattern: str,
-) -> None:
-    """Assert that multi-source ingestion raises a matching ValueError."""
-    async with SqlAlchemyUnitOfWork(context.session_factory) as uow:
-        with pytest.raises(ValueError, match=expected_error_pattern):
-            await ingest_multi_source(
-                uow,
-                context.profile,
-                request,
-                context.ingestion_pipeline,
-            )
 
 
 @pytest.fixture
@@ -185,200 +26,12 @@ def ingestion_test_context(
     session_factory: typ.Callable[[], AsyncSession],
     series_profile_for_ingestion: SeriesProfile,
     ingestion_pipeline: IngestionPipeline,
-) -> IngestionTestContext:
+) -> ingestion_support.IngestionTestContext:
     """Compose ingestion test fixtures into a context object."""
-    return IngestionTestContext(
+    return ingestion_support.IngestionTestContext(
         session_factory=session_factory,
         profile=series_profile_for_ingestion,
         ingestion_pipeline=ingestion_pipeline,
-    )
-
-
-def _make_reference_ingestion_request(profile: SeriesProfile) -> MultiSourceRequest:
-    """Build the multi-source request used by reference snapshotting tests."""
-    return MultiSourceRequest(
-        raw_sources=[
-            _make_raw_source(
-                source_type="transcript",
-                source_uri="s3://bucket/reference-transcript.txt",
-                content="Primary transcript",
-                content_hash="hash-primary-reference",
-                metadata={"title": "Primary Episode"},
-            ),
-            _make_raw_source(
-                source_type="brief",
-                source_uri="s3://bucket/reference-brief.txt",
-                content="Background brief",
-                content_hash="hash-brief-reference",
-                metadata={"title": "Brief Notes"},
-            ),
-        ],
-        series_slug=profile.slug,
-        requested_by="producer@example.com",
-    )
-
-
-def _assert_reference_snapshot(
-    documents: list[typ.Any],
-    reference_document: ReferenceDocument,
-    reference_revision: ReferenceDocumentRevision,
-    reference_binding: ReferenceBinding,
-) -> None:
-    """Assert that one reference snapshot document was persisted correctly."""
-    snapshot_documents = [
-        document
-        for document in documents
-        if document.source_type == "reference_document"
-    ]
-    assert len(snapshot_documents) == 1, (
-        "Expected exactly one persisted reference snapshot document."
-    )
-    assert len(documents) >= 3, (
-        "Expected at least two raw sources plus one reference snapshot document."
-    )
-    reference_snapshot = snapshot_documents[0]
-    assert reference_snapshot.reference_document_revision_id == reference_revision.id
-    assert reference_snapshot.source_uri == (
-        f"ref://{reference_document.id}/revisions/{reference_revision.id}"
-    )
-    assert reference_snapshot.metadata["binding_id"] == str(reference_binding.id)
-    assert reference_snapshot.metadata["document_kind"] == "style_guide"
-
-
-@pytest.mark.asyncio
-async def test_ingest_multi_source_end_to_end(
-    session_factory: typ.Callable[[], AsyncSession],
-    series_profile_for_ingestion: SeriesProfile,
-    ingestion_pipeline: IngestionPipeline,
-) -> None:
-    """End-to-end integration test for multi-source ingestion."""
-    profile = series_profile_for_ingestion
-
-    request = MultiSourceRequest(
-        raw_sources=[
-            _make_raw_source(
-                source_type="transcript",
-                source_uri="s3://bucket/transcript.txt",
-                content="Primary transcript",
-                content_hash="hash-primary",
-                metadata={"title": "Primary Episode"},
-            ),
-            _make_raw_source(
-                source_type="brief",
-                source_uri="s3://bucket/brief.txt",
-                content="Background brief",
-                content_hash="hash-brief",
-                metadata={"title": "Brief Notes"},
-            ),
-        ],
-        series_slug=profile.slug,
-        requested_by="producer@example.com",
-    )
-
-    async with SqlAlchemyUnitOfWork(session_factory) as uow:
-        episode = await ingest_multi_source(
-            uow,
-            profile,
-            request,
-            ingestion_pipeline,
-        )
-
-    async with SqlAlchemyUnitOfWork(session_factory) as uow:
-        persisted = await uow.episodes.get(episode.id)
-        assert persisted is not None, (
-            "Expected persisted episode to be retrievable after ingestion."
-        )
-        header = await uow.tei_headers.get(persisted.tei_header_id)
-
-    assert persisted.title == "Primary Episode", (
-        "Expected winning source title to persist as canonical episode title."
-    )
-    assert_uuid7(persisted.id, "canonical episode")
-    assert_uuid7(persisted.tei_header_id, "TEI header reference")
-    assert header is not None, "Expected a persisted TEI header."
-    assert_uuid7(header.id, "TEI header")
-
-    provenance = _require_provenance_payload(header.payload)
-    _verify_provenance_metadata(
-        provenance,
-        "producer@example.com",
-        [
-            "s3://bucket/transcript.txt",
-            "s3://bucket/brief.txt",
-        ],
-    )
-
-    job_record = await _get_job_record_for_episode(session_factory, episode.id)
-    assert_uuid7(job_record.id, "ingestion job")
-
-    async with SqlAlchemyUnitOfWork(session_factory) as uow:
-        documents = await uow.source_documents.list_for_job(job_record.id)
-
-    _verify_source_documents(documents, expected_count=2)
-
-
-@pytest.mark.asyncio
-async def test_ingest_multi_source_preserves_all_sources(
-    session_factory: typ.Callable[[], AsyncSession],
-    series_profile_for_ingestion: SeriesProfile,
-    ingestion_pipeline: IngestionPipeline,
-) -> None:
-    """All sources are persisted, even those rejected in conflict resolution."""
-    profile = series_profile_for_ingestion
-
-    source_uris = [
-        "s3://bucket/source-1.txt",
-        "s3://bucket/source-2.txt",
-        "s3://bucket/source-3.txt",
-    ]
-    request = MultiSourceRequest(
-        raw_sources=[
-            _make_raw_source(
-                source_type="transcript",
-                source_uri=source_uris[0],
-                content="Source one",
-                content_hash="hash-1",
-                metadata={"title": "Source One"},
-            ),
-            _make_raw_source(
-                source_type="brief",
-                source_uri=source_uris[1],
-                content="Source two",
-                content_hash="hash-2",
-                metadata={"title": "Source Two"},
-            ),
-            _make_raw_source(
-                source_type="rss",
-                source_uri=source_uris[2],
-                content="Source three",
-                content_hash="hash-3",
-                metadata={"title": "Source Three"},
-            ),
-        ],
-        series_slug=profile.slug,
-        requested_by="test@example.com",
-    )
-
-    async with SqlAlchemyUnitOfWork(session_factory) as uow:
-        episode = await ingest_multi_source(
-            uow,
-            profile,
-            request,
-            ingestion_pipeline,
-        )
-
-    job_record = await _get_job_record_for_episode(
-        session_factory,
-        episode.id,
-    )
-
-    async with SqlAlchemyUnitOfWork(session_factory) as uow:
-        documents = await uow.source_documents.list_for_job(job_record.id)
-
-    assert len(documents) == 3, "Expected all three sources to be persisted."
-    persisted_uris = {doc.source_uri for doc in documents}
-    assert persisted_uris == set(source_uris), (
-        "Expected persisted source URIs to match all input URIs."
     )
 
 
@@ -400,15 +53,12 @@ async def test_ingest_multi_source_preserves_all_sources(
     ids=["empty_sources", "slug_mismatch"],
 )
 async def test_ingest_multi_source_validation_errors(
-    ingestion_test_context: IngestionTestContext,
+    ingestion_test_context: ingestion_support.IngestionTestContext,
     raw_sources: list,
     series_slug: str | None,
     expected_error_pattern: str,
 ) -> None:
-    """Multi-source ingestion validates input and raises ValueError.
-
-    Invalid requests raise validation errors.
-    """
+    """Multi-source ingestion validates input and raises ValueError."""
     request = MultiSourceRequest(
         raw_sources=raw_sources,
         series_slug=(
@@ -418,119 +68,8 @@ async def test_ingest_multi_source_validation_errors(
         ),
         requested_by="test@example.com",
     )
-    await _assert_ingestion_raises(
+    await ingestion_support.assert_ingestion_raises(
         ingestion_test_context,
         request,
         expected_error_pattern,
-    )
-
-
-@pytest.mark.asyncio
-async def test_ingest_multi_source_records_conflict_metadata(
-    session_factory: typ.Callable[[], AsyncSession],
-    series_profile_for_ingestion: SeriesProfile,
-    ingestion_pipeline: IngestionPipeline,
-) -> None:
-    """Conflict-resolution metadata is recorded in source document metadata."""
-    profile = series_profile_for_ingestion
-
-    request = MultiSourceRequest(
-        raw_sources=[
-            _make_raw_source(
-                source_type="transcript",
-                source_uri="s3://bucket/winner.txt",
-                content="Winner content",
-                content_hash="hash-winner",
-                metadata={"title": "Winner"},
-            ),
-            _make_raw_source(
-                source_type="rss",
-                source_uri="s3://bucket/loser.txt",
-                content="Loser content",
-                content_hash="hash-loser",
-                metadata={"title": "Loser"},
-            ),
-        ],
-        series_slug=profile.slug,
-        requested_by="test@example.com",
-    )
-
-    async with SqlAlchemyUnitOfWork(session_factory) as uow:
-        episode = await ingest_multi_source(
-            uow,
-            profile,
-            request,
-            ingestion_pipeline,
-        )
-        persisted = await uow.episodes.get(episode.id)
-        assert persisted is not None, "Expected persisted canonical episode."
-        header = await uow.tei_headers.get(persisted.tei_header_id)
-        assert header is not None, "Expected persisted TEI header."
-
-    job_record = await _get_job_record_for_episode(
-        session_factory,
-        episode.id,
-    )
-
-    async with SqlAlchemyUnitOfWork(session_factory) as uow:
-        documents = await uow.source_documents.list_for_job(job_record.id)
-
-    for doc in documents:
-        assert "conflict_resolution" in doc.metadata, (
-            "Expected conflict-resolution metadata to be attached to each source."
-        )
-        cr = typ.cast("dict[str, object]", doc.metadata["conflict_resolution"])
-        assert "preferred_sources" in cr, (
-            "Expected conflict metadata to include preferred sources."
-        )
-        assert "rejected_sources" in cr, (
-            "Expected conflict metadata to include rejected sources."
-        )
-        assert "resolution_notes" in cr, (
-            "Expected conflict metadata to include resolver notes."
-        )
-
-    provenance = _require_provenance_payload(header.payload)
-    priorities = provenance["source_priorities"]
-    assert priorities[0]["source_uri"] == "s3://bucket/winner.txt", (
-        "Expected winner source to lead priority ordering."
-    )
-    assert priorities[1]["source_uri"] == "s3://bucket/loser.txt", (
-        "Expected loser source to follow in priority ordering."
-    )
-
-
-@pytest.mark.asyncio
-async def test_ingest_multi_source_snapshots_resolved_reference_bindings(
-    session_factory: typ.Callable[[], AsyncSession],
-    series_profile_for_ingestion: SeriesProfile,
-    ingestion_pipeline: IngestionPipeline,
-) -> None:
-    """Series-level resolved bindings should be snapshotted as source documents."""
-    profile = series_profile_for_ingestion
-    (
-        reference_document,
-        reference_revision,
-        reference_binding,
-    ) = await _create_reference_fixtures(session_factory, profile)
-    request = _make_reference_ingestion_request(profile)
-
-    async with SqlAlchemyUnitOfWork(session_factory) as uow:
-        episode = await ingest_multi_source(
-            uow,
-            profile,
-            request,
-            ingestion_pipeline,
-        )
-
-    job_record = await _get_job_record_for_episode(session_factory, episode.id)
-
-    async with SqlAlchemyUnitOfWork(session_factory) as uow:
-        documents = await uow.source_documents.list_for_job(job_record.id)
-
-    _assert_reference_snapshot(
-        documents,
-        reference_document,
-        reference_revision,
-        reference_binding,
     )
